// frontend/src/Utils/types.ts
export interface Food {
    _id?: string; // Optional for creation, will be generated by MongoDB
    name: string;
    description: string;
    image?: string;
    sizes: FoodSize[];
    category: string;
    rating?: number; // Should be between 0-5 when provided
    raters?:number;
    available?: boolean;
    veg?:boolean | null,
    createdAt?: Date; // Will be automatically set by the schema
    updatedAt?: Date; // Will be automatically set by the schema
  }
  
  // For more type safety with size options:
  export interface FoodSize {
    name: string;
    price: number;
  }

  export interface Category{
    _id?:string
    name: string;
    icon?: string;
    createdAt: Date;
    updatedAt: Date;
  }


  export interface Item {
    _id: string;
    category?:string,
    name: string;
    price: number;
    qty: number;
    size: string;
  }
  
  interface OrderStatus {
    status: string;
    time: Date;
  }
  
  export interface Order {

    _id: string;
    createdAt: string;
    items: Item[];
    message?: string;
    currentStatus?: OrderStatus | undefined; 
    paymentMethod: string;
    user?: User;
  }


  export interface User{
    _id:string,
    email: string;
    name: string;
    picture?: string;
    password?: string;
    role?: string;
    isActive: boolean;
    isVerified: boolean;
    verifyToken?: string;
    verifyTokenExpiry?: Date;
    forgotPasswordToken?: string;
    forgotPasswordExpiry?: Date;
    createdAt: Date;
    updatedAt: Date;
  }